+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased -- 2023-mm-dd
 
+- **Breaking:** Remove `TypedBuilder` implementation for `server::Config`.
+
 ## v0.1.1 - 2023-07-30
 
-- Functional with the stable version of the Rust compiler
+- Support stable version of the Rust compiler
 
 ## v0.1.0 - 2023-07-29